+++ assets/stylesheets/_components.scss
@@ -247,8 +247,8 @@
 @import 'me/purchases/confirm-cancel-domain/style';
 @import 'me/purchases/credit-cards/credit-card-delete';
 @import 'me/purchases/credit-cards/credit-cards';
-@import 'me/purchases/list/item/style';
-@import 'me/purchases/list/site/style';
+@import 'me/purchases/purchase-item/style';
+@import 'me/purchases/purchases-site/style';
 @import 'me/purchases/manage-purchase/style';
 @import 'me/purchases/remove-purchase/style';
 @import 'me/reauth-required/style';
