+++ client/me/purchases/controller.jsx
@@ -17,8 +17,8 @@ import Main from 'components/main';
 import ManagePurchase from './manage-purchase';
 import NoSitesMessage from 'components/empty-content/no-sites-message';
 import paths from './paths';
-import PurchasesHeader from './list/header';
-import PurchasesList from './list';
+import PurchasesHeader from './purchases-list/header';
+import PurchasesList from './purchases-list';
 import { receiveSite } from 'state/sites/actions';
 import { concatTitle, recordPageView, renderPage } from 'lib/react-helpers';
 import { setAllSitesSelected, setSelectedSiteId } from 'state/ui/actions';
